app/requirements.txt
flask==3.0.0
openai>=1.54.0,<2.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pandas==2.1.4
langchain-openai==0.2.8
